requirements.txt
SQLAlchemy
python-geoip
python-geoip-geolite2
ConfigParser
